--- a/Source/Editor/Surface/Archetypes/Function.cs
+++ b/Source/Editor/Surface/Archetypes/Function.cs
@@ -773,7 +773,7 @@ namespace FlaxEditor.Surface.Archetypes
                     Values[4] = GetSignatureData(memberInfo, memberInfo.GetParameters());
                 }
             }
-
+            
             private SignatureInfo LoadSignature()
             {
                 var signature = new SignatureInfo();
@@ -1150,6 +1150,45 @@ namespace FlaxEditor.Surface.Archetypes
                 }
 
                 base.OnDestroy();
+            }
+            
+            internal static bool IsInputCompatible(NodeArchetype nodeArch, ScriptType outputType, ConnectionsHint hint)
+            {
+                if (nodeArch.Tag is not ScriptMemberInfo memberInfo)
+                    return false;
+                if(memberInfo.ValueType.IsVoid && outputType.IsVoid)
+                    return true;
+                if (!memberInfo.IsStatic)
+                {
+                    if (VisjectSurface.FullCastCheck(memberInfo.DeclaringType, outputType, hint))
+                        return true;
+                }
+                foreach (var param in memberInfo.GetParameters())
+                {
+                    if(param.IsOut)
+                        continue;
+                    if (VisjectSurface.FullCastCheck(param.Type, outputType, hint))
+                        return true;
+                }
+                return false;
+            }
+            
+            internal static bool IsOutputCompatible(NodeArchetype nodeArch, ScriptType inputType, ConnectionsHint hint)
+            {
+                if (nodeArch.Tag is not ScriptMemberInfo memberInfo)
+                    return false;
+                if(memberInfo.ValueType.IsVoid && inputType.IsVoid)
+                    return true;
+                if (VisjectSurface.FullCastCheck(memberInfo.ValueType, inputType, hint))
+                    return true;
+                foreach (var param in memberInfo.GetParameters())
+                {
+                    if(!param.IsOut)
+                        continue;
+                    if (VisjectSurface.FullCastCheck(param.Type, inputType, hint))
+                        return true;
+                }
+                return false;
             }
         }
 
@@ -2228,7 +2267,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Title = "Function Input",
                 Description = "The graph function input data",
                 Flags = NodeFlags.MaterialGraph | NodeFlags.ParticleEmitterGraph | NodeFlags.AnimGraph | NodeFlags.NoSpawnViaPaste,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 60),
                 DefaultValues = new object[]
                 {
@@ -2248,7 +2286,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Title = "Function Output",
                 Description = "The graph function output data",
                 Flags = NodeFlags.MaterialGraph | NodeFlags.ParticleEmitterGraph | NodeFlags.AnimGraph | NodeFlags.NoSpawnViaPaste,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 60),
                 DefaultValues = new object[]
                 {
@@ -2267,7 +2304,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Title = string.Empty,
                 Description = "Overrides the base class method with custom implementation",
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI | NodeFlags.NoSpawnViaPaste,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 60),
                 DefaultValues = new object[]
                 {
@@ -2280,9 +2316,10 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 4,
                 Create = (id, context, arch, groupArch) => new InvokeMethodNode(id, context, arch, groupArch),
+                IsInputCompatible = InvokeMethodNode.IsInputCompatible,
+                IsOutputCompatible = InvokeMethodNode.IsOutputCompatible,
                 Title = string.Empty,
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 60),
                 DefaultValues = new object[]
                 {
@@ -2306,7 +2343,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Create = (id, context, arch, groupArch) => new ReturnNode(id, context, arch, groupArch),
                 Title = "Return",
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(100, 40),
                 DefaultValues = new object[]
                 {
@@ -2325,7 +2361,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Title = "New Function",
                 Description = "Adds a new function to the script",
                 Flags = NodeFlags.VisualScriptGraph,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 20),
                 DefaultValues = new object[]
                 {
@@ -2338,7 +2373,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Create = (id, context, arch, groupArch) => new GetFieldNode(id, context, arch, groupArch),
                 Title = string.Empty,
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 60),
                 DefaultValues = new object[]
                 {
@@ -2354,7 +2388,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Create = (id, context, arch, groupArch) => new SetFieldNode(id, context, arch, groupArch),
                 Title = string.Empty,
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(240, 60),
                 DefaultValues = new object[]
                 {
@@ -2371,7 +2404,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Create = (id, context, arch, groupArch) => new BindEventNode(id, context, arch, groupArch),
                 Title = string.Empty,
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(260, 60),
                 DefaultValues = new object[]
                 {
@@ -2394,7 +2426,6 @@ namespace FlaxEditor.Surface.Archetypes
                 Create = (id, context, arch, groupArch) => new UnbindEventNode(id, context, arch, groupArch),
                 Title = string.Empty,
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.NoSpawnViaGUI,
-                NodeTypeHint = NodeTypeHint.FunctionNode,
                 Size = new Float2(260, 60),
                 DefaultValues = new object[]
                 {