When dragging a connection out of a node port into the void, the context menu to add nodes gets shown.
Context sensitive node search makes it so that only nodes with compatible ports get shown in the context menu. This mimics the behaviour of visual editors from other engines, e.g. Unreal.


### Tasks
- [x] Implement context menu item filtering by type
- [x] Implement context menu group filtering by type
- [x] Support default archetypes and groups
- [x] Support particle emitter graph specific archetypes and groups
- [x] Support visual scripting specific archetypes and groups 
- [x] Support animation graph specific archetypes and groups
- [x] Add "context sensitive" toggle to context menu
- [x] Cleanup
- [x] Add XML comments
- [x] A lot of testing 
- [x] New tasks ahead
- [x] Rewrite input/output compatibility check for custom nodes
- [x] Cleanup 2
- [x] Testing 2
- [x] Profiling (against master)
- [x] Done 🥳
- [x] #1114 

**Known Bugs**
- [x] 'Comment' archetype shows up when typing in the search bar
- [x] Filtering by group name activates all items (turned off at the moment)
- [x] Asynchronously loaded nodes that aren't InvokeMethodNodes currently get filtered out even when compatible

**Optional**
- [x] Make context menu taller and less wide
- [x] Make item entries a tiny bit taller

**Didn't do**
- [x] Change entry score based on the port type (e.g. when dragging out of a vector3 port, options that have Vector3 as input should have a higher score than Float3, Vector2, int etc) -> Will probably be its own PR

### Profiling Results
I didn't do craaaazzy profiling and testing. But enough to get an idea what the costs of this feature are.
I used the in engine profiler to check how much longer in time the opening and fetching of the items take, and how much more memory is getting used. Also these times are just the spikes visible in the profiler, since custom nodes get fetched asynchronously, so the time for all items to appear varies.

First of all, when right clicking into the void, so when no context is really needed, the speeds are rather identical. 
Opening the CM took in both cases around **85ms** on average while taking up **~4MB of memory**

Now, the more interesting part: dragging a connection out of a port. Since here context sensitivity really just starts doing work.

On master with no context sensitivity it took around **85ms** on average while taking up **~4MB of memory**, what a surprise
And on my branch with context sensitivity enabled it took around **116ms** while taking up **~7MB of memory**

I think it comes to no surprise that filtering through all archetypes does take some more time. With an average of around 30ms more, which means it's effectively ~35% slower now and takes up 3MB of memory more.

Like i said, i didn't do crazy profiling here, i just read numbers from the in engine profiler. Let's just say i was a bit lazy to open dotTrace now and do in depth profiling.
Also the whole filtering process could probably be a lot more optimized and faster, but would need some refactoring.
I think sacrificing a couple of milliseconds more for such a QoL feature is okay, when opening a context menu. Especially since it can still be optimized.

Just keep in mind that my PC is on the stronger side, with an Ryzen 9 3900X, 32GB RAM etc. But i think even weaker PCs won't have a problem with that.

### Feature preview

Material Editor
![ContextualNodes](https://github.com/FlaxEngine/FlaxEngine/assets/6757167/2680cc8f-a996-49ca-a541-79befe9a1db7)

Visual Scripting Editor
![ContextSensitiveVisualScripting](https://github.com/FlaxEngine/FlaxEngine/assets/6757167/3218aa5c-042e-49bc-b89d-b8103596bc42)

Context Menu Toggle
![ContextSensitiveToggle](https://github.com/FlaxEngine/FlaxEngine/assets/6757167/3dc6bc2d-f39f-407f-ba3e-e472629761c6)