@Chikinsupu 2023-10-01
>@Chikinsupu How ready is this? Do you need another testing week or can we merge this week?
>
>@stefnotch Would you mind taking a look and review?

I tested all kinds of graphs. Material, Particle, Visual Scripting and Animation and i didn't encounter any more issues after my last tests. So i would say it does work generally, or at least it won't break anything.
If needed i can test for another week of course and check for edge cases.

We just have to keep in mind that this is a rather stupid context filter, since all it does as of now, is to check if an archetype is compatible with the current port type. No more, no less.

I think a more interesting question would be, if this is sufficient for now and can be updated over time, or if this feature needs some additional functionality.
Examples:

- Right now it doesn't update item scores, meaning that nodes that would make more sense in the given context get pushed up the list. So, right now when you have a port with a Vector3 type, and drag a connection out of there, it would still show float2, float3, float4, double2, double3, double4, vector2 methods before the vector3 methods.

- Automatic casting is not implemented. Right now a ton of nodes are valid because the underlying system thinks that a type can be casted to another type. When you drag a connection out of an 'Actor' port, it shows you a ton of items that will not connect after spawning the node, because the Actor needs to be casted to another type first. This was already commented in the code, prior to this PR, just not implemented and instead it throws an exception right now.

So, depending on if this PR should fix things like the 2 mentioned examples, it's ready or not.
I was thinking of making 2 different PRs for automatic casting, and smarter sorting, but fixing those things in this PR is fine for me as well.

Any thoughts?