@stefnotch 2023-10-01
I only took a quick glance at the code and haven't thoroughly tested this. So here are my preliminary 2 cents:
The code does look well written, finding bugs would probably be a matter of testing it.
In terms of features, it's definitely sufficient for now. I'd love to see the smarter sorting and automatic casting being implemented in future PRs.